pi/requirements.txt
paho-mqtt==1.6.1
python-espeak==0.5
